Now we are manually specifying red, green, and blue color offsets
In the previous commit, we were setting the color offsets to be the same for the r, g, and b channels (r = red, g = green, b = blue). This meant that the color offsets always ended up being shades of grey.

In this commit we changed things so that the color offets are now r = 0, g = 1, b = 0. This means that as the y value of our terrain increases it goes from black to green instead of black to white.