Added a rule to color areas near the water's edge brown
This rule deserves a bit of explanation:

To create the proper brown gradient we need to go from light brown to dark brown as y increases. This is opposite to how we generate the green gradien
t. The green gradient goes from dark (at low y values) to light (at higher y values).

To get the gradient going in the opposite direction we need we subtract double the upper limit (0.05 in this case) of our rule from `p`, which will be
 a negative value. As `p` get closer to the upper limit `r` will become smaller - meaning the `r` and `g` channel values get smaller (meaning they get darker).